+++ src/pages/Login.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { useRef } from 'react';
 import { toast } from 'react-toastify';
 import { json, useNavigate } from 'react-router-dom';
+import { login } from '../services';
 
 export const Login = () => {
 
@@ -14,22 +15,9 @@ export const Login = () => {
       email: email.current.value,
       password :password.current.value
     }
-
-    const requestOptions = {
-      method : "POST",
-      headers: {"Content-Type": "application/json"},
-      body: JSON.stringify(authDetail)
-    }
-    const response = await fetch("http://localhost:8000/login", requestOptions);
-    const data = await response.json();
-    console.log(data);
+    const data = await login(authDetail);
     data.accessToken ? navigate("/products") : toast.error(data);
-
-
-    if(data.accessToken){
-      sessionStorage.setItem("token", JSON.stringify(data.accessToken));
-      sessionStorage.setItem("cbid", JSON.stringify(data.user.id));
-    }
+  
 
 
   }
