+++ src/pages/Register.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { toast } from 'react-toastify';
 import { useNavigate } from 'react-router-dom';
+import { register } from '../services/authService';
 
 export const Register = () => {
 
@@ -14,19 +15,8 @@ export const Register = () => {
       password: event.target.password.value
     }
 
-    const requestOptions = {
-      method : "POST",
-      headers: {"Content-Type": "application/json"},
-      body: JSON.stringify(authDetail)
-    }
-     const response = await fetch("http://localhost:8000/register", requestOptions);
-     const data = await response.json();
-     data.accessToken ? navigate("/products") : toast.error(data);
-     console.log(data);
-     if(data.accessToken){
-      sessionStorage.setItem("token", JSON.stringify(data.accessToken));
-      sessionStorage.setItem("cbid", JSON.stringify(data.user.id));
-    }
+    const data = await register(authDetail);
+    data.accessToken ? navigate("/products") : toast.error(data);
 
 
    }
