+++ src/services/authService.js
@@ -1,0 +1,41 @@
+
+export async function login(authDetail) {
+    const requestOptions = {
+        method : "POST",
+        headers: {"Content-Type": "application/json"},
+        body: JSON.stringify(authDetail)
+      }
+      const response = await fetch("http://localhost:8000/login", requestOptions);
+      const data = await response.json();
+      console.log(data);
+   
+  
+  
+      if(data.accessToken){
+        sessionStorage.setItem("token", JSON.stringify(data.accessToken));
+        sessionStorage.setItem("cbid", JSON.stringify(data.user.id));
+      }
+    return data;
+}
+
+export async function register(authDetail) {
+  const requestOptions = {
+    method : "POST",
+    headers: {"Content-Type": "application/json"},
+    body: JSON.stringify(authDetail)
+  }
+   const response = await fetch("http://localhost:8000/register", requestOptions);
+   const data = await response.json();
+   console.log(data);
+   if(data.accessToken){
+    sessionStorage.setItem("token", JSON.stringify(data.accessToken));
+    sessionStorage.setItem("cbid", JSON.stringify(data.user.id));
+  }
+  return data;
+}
+
+export async function logout(authDetail) {
+  sessionStorage.removeItem("token");
+  sessionStorage.removeItem("cbid");
+
+}
